Move and organize AST code; extend smart printing.
