+++ app/gui2/package.json
@@ -24,8 +24,8 @@
     "lint": "eslint .",
     "format": "prettier --version && prettier --write src/ && eslint . --fix",
     "build-rust-ffi": "wasm-pack build ./rust-ffi --release --target web",
-    "generate-ast-schema": "cargo run -p enso-parser-schema > src/generated/ast-schema.json",
-    "generate-ast-types": "tsx ./parser-codegen/index.ts src/generated/ast-schema.json src/generated/ast.ts",
+    "generate-ast-schema": "cargo run -p enso-parser-schema > shared/ast/generated/ast-schema.json",
+    "generate-ast-types": "tsx ./parser-codegen/index.ts shared/ast/generated/ast-schema.json shared/ast/generated/ast.ts",
     "preinstall": "npm run build-rust-ffi && npm run generate-ast-schema && npm run generate-ast-types && npm run generate-metadata && npm run download-fonts",
     "postinstall": "playwright install",
     "generate-metadata": "node scripts/generateIconMetadata.js",
