+++ app/gui2/shared/ast/parse.ts
@@ -1,0 +1,408 @@
+import * as map from 'lib0/map'
+import { assert } from '../util/assert'
+import type { SourceRange, SourceRangeKey } from '../yjsModel.ts'
+import { IdMap, isUuid, sourceRangeFromKey, sourceRangeKey } from '../yjsModel.ts'
+import { parse_tree } from './ffi.ts'
+import * as RawAst from './generated/ast'
+import type { AstId, NodeChild, Owned } from './index.ts'
+import { Token, asOwned, parentId } from './index.ts'
+import { MutableModule } from './mutableModule.ts'
+import type { LazyObject } from './parserSupport'
+import {
+  App,
+  Assignment,
+  Ast,
+  BodyBlock,
+  Documented,
+  Function,
+  Generic,
+  Group,
+  Ident,
+  Import,
+  Invalid,
+  MutableBodyBlock,
+  MutableIdent,
+  NegationApp,
+  NumericLiteral,
+  OprApp,
+  PropertyAccess,
+  TextLiteral,
+  UnaryOprApp,
+  Wildcard,
+} from './tree.ts'
+
+const DEBUG = false
+
+export function parseEnso(code: string): RawAst.Tree {
+  const blob = parse_tree(code)
+  return RawAst.Tree.read(new DataView(blob.buffer), blob.byteLength - 4)
+}
+
+export function normalize(rootIn: Ast): Ast {
+  const printed = print(rootIn)
+  const idMap = spanMapToIdMap(printed.info)
+  const module = MutableModule.Transient()
+  const tree = parseEnso(printed.code)
+  const { root: parsed, spans } = abstract(module, tree, printed.code)
+  module.replaceRoot(parsed)
+  setExternalIds(module, spans, idMap)
+  return parsed
+}
+
+export function abstract(
+  module: MutableModule,
+  tree: RawAst.Tree,
+  code: string,
+): { root: Owned; spans: SpanMap; toRaw: Map<AstId, RawAst.Tree> } {
+  const tokens = new Map()
+  const nodes = new Map()
+  const toRaw = new Map()
+  const root = abstractTree(module, tree, code, nodes, tokens, toRaw).node
+  const spans = { tokens, nodes }
+  return { root, spans, toRaw }
+}
+
+function abstractTree(
+  module: MutableModule,
+  tree: RawAst.Tree,
+  code: string,
+  nodesOut: NodeSpanMap,
+  tokensOut: TokenSpanMap,
+  toRaw: Map<AstId, RawAst.Tree>,
+): { whitespace: string | undefined; node: Owned } {
+  const recurseTree = (tree: RawAst.Tree) =>
+    abstractTree(module, tree, code, nodesOut, tokensOut, toRaw)
+  const recurseToken = (token: RawAst.Token.Token) => abstractToken(token, code, tokensOut)
+  const visitChildren = (tree: LazyObject) => {
+    const children: NodeChild<Owned | Token>[] = []
+    const visitor = (child: LazyObject) => {
+      if (RawAst.Tree.isInstance(child)) {
+        children.push(recurseTree(child))
+      } else if (RawAst.Token.isInstance(child)) {
+        children.push(recurseToken(child))
+      } else {
+        child.visitChildren(visitor)
+      }
+    }
+    tree.visitChildren(visitor)
+    return children
+  }
+  const whitespaceStart = tree.whitespaceStartInCodeParsed
+  const whitespaceEnd = whitespaceStart + tree.whitespaceLengthInCodeParsed
+  const whitespace = code.substring(whitespaceStart, whitespaceEnd)
+  const codeStart = whitespaceEnd
+  const codeEnd = codeStart + tree.childrenLengthInCodeParsed
+  const spanKey = nodeKey(codeStart, codeEnd - codeStart)
+  let node: Owned
+  switch (tree.type) {
+    case RawAst.Tree.Type.BodyBlock: {
+      const lines = Array.from(tree.statements, (line) => {
+        const newline = recurseToken(line.newline)
+        const expression = line.expression ? recurseTree(line.expression) : undefined
+        return { newline, expression }
+      })
+      node = BodyBlock.concrete(module, lines)
+      break
+    }
+    case RawAst.Tree.Type.Function: {
+      const name = recurseTree(tree.name)
+      const argumentDefinitions = Array.from(tree.args, (arg) => visitChildren(arg))
+      const equals = recurseToken(tree.equals)
+      const body = tree.body !== undefined ? recurseTree(tree.body) : undefined
+      node = Function.concrete(module, name, argumentDefinitions, equals, body)
+      break
+    }
+    case RawAst.Tree.Type.Ident: {
+      const token = recurseToken(tree.token)
+      node = Ident.concrete(module, token)
+      break
+    }
+    case RawAst.Tree.Type.Assignment: {
+      const pattern = recurseTree(tree.pattern)
+      const equals = recurseToken(tree.equals)
+      const value = recurseTree(tree.expr)
+      node = Assignment.concrete(module, pattern, equals, value)
+      break
+    }
+    case RawAst.Tree.Type.App: {
+      const func = recurseTree(tree.func)
+      const arg = recurseTree(tree.arg)
+      node = App.concrete(module, func, undefined, undefined, arg)
+      break
+    }
+    case RawAst.Tree.Type.NamedApp: {
+      const func = recurseTree(tree.func)
+      const open = tree.open ? recurseToken(tree.open) : undefined
+      const name = recurseToken(tree.name)
+      const equals = recurseToken(tree.equals)
+      const arg = recurseTree(tree.arg)
+      const close = tree.close ? recurseToken(tree.close) : undefined
+      const parens = open && close ? { open, close } : undefined
+      const nameSpecification = { name, equals }
+      node = App.concrete(module, func, parens, nameSpecification, arg)
+      break
+    }
+    case RawAst.Tree.Type.UnaryOprApp: {
+      const opr = recurseToken(tree.opr)
+      const arg = tree.rhs ? recurseTree(tree.rhs) : undefined
+      if (arg && opr.node.code() === '-') {
+        node = NegationApp.concrete(module, opr, arg)
+      } else {
+        node = UnaryOprApp.concrete(module, opr, arg)
+      }
+      break
+    }
+    case RawAst.Tree.Type.OprApp: {
+      const lhs = tree.lhs ? recurseTree(tree.lhs) : undefined
+      const opr = tree.opr.ok
+        ? [recurseToken(tree.opr.value)]
+        : Array.from(tree.opr.error.payload.operators, recurseToken)
+      const rhs = tree.rhs ? recurseTree(tree.rhs) : undefined
+      if (opr.length === 1 && opr[0]?.node.code() === '.' && rhs?.node instanceof MutableIdent) {
+        // Propagate type.
+        const rhs_ = { ...rhs, node: rhs.node }
+        node = PropertyAccess.concrete(module, lhs, opr[0], rhs_)
+      } else {
+        node = OprApp.concrete(module, lhs, opr, rhs)
+      }
+      break
+    }
+    case RawAst.Tree.Type.Number: {
+      const tokens = []
+      if (tree.base) tokens.push(recurseToken(tree.base))
+      if (tree.integer) tokens.push(recurseToken(tree.integer))
+      if (tree.fractionalDigits) {
+        tokens.push(recurseToken(tree.fractionalDigits.dot))
+        tokens.push(recurseToken(tree.fractionalDigits.digits))
+      }
+      node = NumericLiteral.concrete(module, tokens)
+      break
+    }
+    case RawAst.Tree.Type.Wildcard: {
+      const token = recurseToken(tree.token)
+      node = Wildcard.concrete(module, token)
+      break
+    }
+    // These expression types are (or will be) used for backend analysis.
+    // The frontend can ignore them, avoiding some problems with expressions sharing spans
+    // (which makes it impossible to give them unique IDs in the current IdMap format).
+    case RawAst.Tree.Type.OprSectionBoundary:
+    case RawAst.Tree.Type.TemplateFunction:
+      return { whitespace, node: recurseTree(tree.ast).node }
+    case RawAst.Tree.Type.Invalid: {
+      const expression = recurseTree(tree.ast)
+      node = Invalid.concrete(module, expression)
+      break
+    }
+    case RawAst.Tree.Type.Group: {
+      const open = tree.open ? recurseToken(tree.open) : undefined
+      const expression = tree.body ? recurseTree(tree.body) : undefined
+      const close = tree.close ? recurseToken(tree.close) : undefined
+      node = Group.concrete(module, open, expression, close)
+      break
+    }
+    case RawAst.Tree.Type.TextLiteral: {
+      const open = tree.open ? recurseToken(tree.open) : undefined
+      const newline = tree.newline ? recurseToken(tree.newline) : undefined
+      const elements = []
+      for (const e of tree.elements) {
+        elements.push(...visitChildren(e))
+      }
+      const close = tree.close ? recurseToken(tree.close) : undefined
+      node = TextLiteral.concrete(module, open, newline, elements, close)
+      break
+    }
+    case RawAst.Tree.Type.Documented: {
+      const open = recurseToken(tree.documentation.open)
+      const elements = []
+      for (const e of tree.documentation.elements) {
+        elements.push(...visitChildren(e))
+      }
+      const newlines = Array.from(tree.documentation.newlines, recurseToken)
+      const expression = tree.expression ? recurseTree(tree.expression) : undefined
+      node = Documented.concrete(module, open, elements, newlines, expression)
+      break
+    }
+    case RawAst.Tree.Type.Import: {
+      const recurseBody = (tree: RawAst.Tree) => {
+        const body = recurseTree(tree)
+        if (body.node instanceof Invalid && body.node.code() === '') return undefined
+        return body
+      }
+      const recurseSegment = (segment: RawAst.MultiSegmentAppSegment) => ({
+        header: recurseToken(segment.header),
+        body: segment.body ? recurseBody(segment.body) : undefined,
+      })
+      const polyglot = tree.polyglot ? recurseSegment(tree.polyglot) : undefined
+      const from = tree.from ? recurseSegment(tree.from) : undefined
+      const import_ = recurseSegment(tree.import)
+      const all = tree.all ? recurseToken(tree.all) : undefined
+      const as = tree.as ? recurseSegment(tree.as) : undefined
+      const hiding = tree.hiding ? recurseSegment(tree.hiding) : undefined
+      node = Import.concrete(module, polyglot, from, import_, all, as, hiding)
+      break
+    }
+    default: {
+      node = Generic.concrete(module, visitChildren(tree))
+    }
+  }
+  toRaw.set(node.id, tree)
+  map.setIfUndefined(nodesOut, spanKey, (): Ast[] => []).push(node)
+  return { node, whitespace }
+}
+
+function abstractToken(
+  token: RawAst.Token,
+  code: string,
+  tokensOut: TokenSpanMap,
+): { whitespace: string; node: Token } {
+  const whitespaceStart = token.whitespaceStartInCodeBuffer
+  const whitespaceEnd = whitespaceStart + token.whitespaceLengthInCodeBuffer
+  const whitespace = code.substring(whitespaceStart, whitespaceEnd)
+  const codeStart = token.startInCodeBuffer
+  const codeEnd = codeStart + token.lengthInCodeBuffer
+  const tokenCode = code.substring(codeStart, codeEnd)
+  const key = tokenKey(codeStart, codeEnd - codeStart)
+  const node = Token.new(tokenCode, token.type)
+  tokensOut.set(key, node)
+  return { whitespace, node }
+}
+
+declare const nodeKeyBrand: unique symbol
+export type NodeKey = SourceRangeKey & { [nodeKeyBrand]: never }
+declare const tokenKeyBrand: unique symbol
+export type TokenKey = SourceRangeKey & { [tokenKeyBrand]: never }
+export function nodeKey(start: number, length: number): NodeKey {
+  return sourceRangeKey([start, start + length]) as NodeKey
+}
+export function tokenKey(start: number, length: number): TokenKey {
+  return sourceRangeKey([start, start + length]) as TokenKey
+}
+
+export type NodeSpanMap = Map<NodeKey, Ast[]>
+export type TokenSpanMap = Map<TokenKey, Token>
+
+export interface SpanMap {
+  nodes: NodeSpanMap
+  tokens: TokenSpanMap
+}
+
+interface PrintedSource {
+  info: SpanMap
+  code: string
+}
+
+export function spanMapToIdMap(spans: SpanMap): IdMap {
+  const idMap = new IdMap()
+  for (const [key, token] of spans.tokens.entries()) {
+    assert(isUuid(token.id))
+    idMap.insertKnownId(sourceRangeFromKey(key), token.id)
+  }
+  for (const [key, asts] of spans.nodes.entries()) {
+    for (const ast of asts) {
+      assert(isUuid(ast.externalId))
+      idMap.insertKnownId(sourceRangeFromKey(key), ast.externalId)
+    }
+  }
+  return idMap
+}
+
+function spanMapToSpanGetter(spans: SpanMap): (id: AstId) => SourceRange | undefined {
+  const reverseMap = new Map<AstId, SourceRange>()
+  for (const [key, asts] of spans.nodes) {
+    for (const ast of asts) {
+      reverseMap.set(ast.id, sourceRangeFromKey(key))
+    }
+  }
+  return (id) => reverseMap.get(id)
+}
+
+/** Return stringification with associated ID map. This is only exported for testing. */
+export function print(ast: Ast): PrintedSource {
+  const info: SpanMap = {
+    nodes: new Map(),
+    tokens: new Map(),
+  }
+  const code = ast.printSubtree(info, 0, undefined)
+  return { info, code }
+}
+
+/** Parse the input as a block. */
+export function parseBlock(code: string, inModule?: MutableModule) {
+  return parseBlockWithSpans(code, inModule).root
+}
+
+/** Parse the input. If it contains a single expression at the top level, return it; otherwise, return a block. */
+export function parse(code: string, module?: MutableModule): Owned {
+  const module_ = module ?? MutableModule.Transient()
+  const ast = parseBlock(code, module_)
+  const [expr] = ast.statements()
+  if (!expr) return ast
+  const parent = parentId(expr)
+  if (parent) module_.delete(parent)
+  expr.fields.set('parent', undefined)
+  return asOwned(expr)
+}
+
+export function parseBlockWithSpans(
+  code: string,
+  inModule?: MutableModule,
+): { root: Owned<MutableBodyBlock>; spans: SpanMap } {
+  const tree = parseEnso(code)
+  const module = inModule ?? MutableModule.Transient()
+  return fromRaw(tree, code, module)
+}
+
+function fromRaw(
+  tree: RawAst.Tree,
+  code: string,
+  inModule?: MutableModule,
+): {
+  root: Owned<MutableBodyBlock>
+  spans: SpanMap
+  toRaw: Map<AstId, RawAst.Tree>
+} {
+  const module = inModule ?? MutableModule.Transient()
+  const ast = abstract(module, tree, code)
+  const spans = ast.spans
+  // The root of the tree produced by the parser is always a `BodyBlock`.
+  const root = ast.root as Owned<MutableBodyBlock>
+  return { root, spans, toRaw: ast.toRaw }
+}
+
+export function parseExtended(code: string, idMap?: IdMap | undefined, inModule?: MutableModule) {
+  const rawRoot = parseEnso(code)
+  const module = inModule ?? MutableModule.Transient()
+  const { root, spans, toRaw, idMapUpdates } = module.ydoc.transact(() => {
+    const { root, spans, toRaw } = fromRaw(rawRoot, code, module)
+    root.module.replaceRoot(root)
+    const idMapUpdates = idMap ? setExternalIds(root.module, spans, idMap) : 0
+    return { root, spans, toRaw, idMapUpdates }
+  })
+  const getSpan = spanMapToSpanGetter(spans)
+  const idMapOut = spanMapToIdMap(spans)
+  return { root, idMap: idMapOut, getSpan, toRaw, idMapUpdates }
+}
+
+export function setExternalIds(module: MutableModule, spans: SpanMap, ids: IdMap) {
+  let astsMatched = 0
+  let idsUnmatched = 0
+  let asts = 0
+  module.root()?.visitRecursiveAst((_ast) => (asts += 1))
+  for (const [key, externalId] of ids.entries()) {
+    const asts = spans.nodes.get(key as NodeKey)
+    if (asts) {
+      for (const ast of asts) {
+        astsMatched += 1
+        module.getVersion(ast).setExternalId(externalId)
+      }
+    } else {
+      idsUnmatched += 1
+    }
+  }
+  if (DEBUG)
+    console.info(
+      `asts=${asts}, astsMatched=${astsMatched}, idsUnmatched=${idsUnmatched}, haveRoot=${!!module.root()}`,
+    )
+  return module.root() ? asts - astsMatched : 0
+}
