+++ app/gui2/src/util/ast/aliasAnalysis.ts
@@ -8,8 +8,8 @@ import {
   readTokenSpan,
 } from '@/util/ast'
 import { MappedKeyMap, MappedSet, NonEmptyStack } from '@/util/containers'
-import type { LazyObject } from '@/util/parserSupport'
 import { rangeIsBefore, sourceRangeKey, type SourceRange } from 'shared/yjsModel'
+import type { LazyObject } from '../../../shared/ast/parserSupport.ts'
 
 const ACCESSOR_OPERATOR = '.'
