+++ app/gui2/src/util/ast/extended.ts
@@ -1,5 +1,3 @@
-import * as Ast from '@/generated/ast'
-import { Token, Tree } from '@/generated/ast'
 import { assert } from '@/util/assert'
 import {
   childrenAstNodesOrTokens,
@@ -14,6 +12,8 @@ import type { Opt } from '@/util/data/opt'
 import * as encoding from 'lib0/encoding'
 import * as sha256 from 'lib0/hash/sha256'
 import * as map from 'lib0/map'
+import * as Ast from 'shared/ast/generated/ast'
+import { Token, Tree } from 'shared/ast/generated/ast'
 import type { ExternalId, IdMap, SourceRange } from 'shared/yjsModel'
 import { markRaw } from 'vue'
