+++ app/gui2/src/util/ast/index.ts
@@ -1,28 +1,23 @@
-import * as RawAst from '@/generated/ast'
 import { assert } from '@/util/assert'
 import * as Ast from '@/util/ast/abstract'
+import { parseEnso } from '@/util/ast/abstract'
 import { AstExtended as RawAstExtended } from '@/util/ast/extended'
 import { isResult, mapOk } from '@/util/data/result'
-import { parse } from '@/util/ffi'
-import { LazyObject, LazySequence } from '@/util/parserSupport'
 import * as map from 'lib0/map'
+import * as RawAst from 'shared/ast/generated/ast'
 import type { SourceRange } from 'shared/yjsModel'
+import { LazyObject, LazySequence } from '../../../shared/ast/parserSupport.ts'
 
-export { Ast, RawAst, RawAstExtended }
+export { Ast, RawAst, RawAstExtended, parseEnso }
 
 export type HasAstRange = SourceRange | RawAst.Tree | RawAst.Token
-
-export function parseEnso(code: string): RawAst.Tree {
-  const blob = parse(code)
-  return RawAst.Tree.read(new DataView(blob.buffer), blob.byteLength - 4)
-}
 
 /** Read a single line of code
  *
  * Is meant to be a helper for tests. If the code is multiline, an exception is raised.
  */
 export function parseEnsoLine(code: string): RawAst.Tree {
-  const block = parseEnso(code)
+  const block = Ast.parseEnso(code)
   assert(block.type === RawAst.Tree.Type.BodyBlock)
   const statements = block.statements[Symbol.iterator]()
   const firstLine = statements.next()
