+++ app/gui2/src/util/docParser.ts
@@ -1,4 +1,4 @@
-import { parse_doc_to_json } from '@/util/ffi'
+import { parse_doc_to_json } from '../../shared/ast/ffi.ts'
 
 export function parseDocs(docs: string): Doc.Section[] {
   const json = parse_doc_to_json(docs)
